firmware: sensor OOR mode and major rework
Add support for the out-of-range mode of the pressure sensor. The idea
is that the sensor could wake the µC and the µC doesn't have to poll the
sensor periodically. Unfortunately, the sensor is only able the maximum
range is just about +/-2.55mbar.

Also document our states, move the state handling into a function.