firmware: sensor: always return mbar
Return sensor readings in mbar. This is the first step to support more
than just one pressure sensor. All sensors will need to return the
measurement in mbar. This will also simplify the code because the MBAR()
and similar macros are not needed anymore. On the flipside, runtime
division is now needed which increases the code size. But these routines
are also needed for the status line printing.